SpecialSpaces

Index

SuiteSplines.SUITESPLINES_PKGS

Base.Matrix

SuiteSplines.AbstractMappings.EvaluationSet

SuiteSplines.AbstractMappings.Field

SuiteSplines.AbstractMappings.GeometricMapping

SuiteSplines.AbstractMappings.ResizableArray

SuiteSplines.BezierBernsteinMethods.BezierSimplex

SuiteSplines.BezierBernsteinMethods.LinearIndex

SuiteSplines.BezierBernsteinMethods.LinearIndex

SuiteSplines.BezierBernsteinMethods.MultiIndices

SuiteSplines.BezierBernsteinMethods.Simplex

SuiteSplines.IgaBase.AbstractMapping

SuiteSplines.IgaBase.AbstractSpline

SuiteSplines.IgaBase.Boundary

SuiteSplines.IgaBase.Degree

SuiteSplines.IgaBase.Dimension

SuiteSplines.IgaBase.QuadratureRule

SuiteSplines.IgaBase.QuadratureRule

SuiteSplines.IgaBase.QuadratureRule

SuiteSplines.IgaBase.Regularity

SuiteSplines.IgaBase.ScalarMapping

SuiteSplines.IgaFormation.Element

SuiteSplines.IgaFormation.ElementAccessor

SuiteSplines.IgaFormation.Elements

SuiteSplines.IgaFormation.PatchAccessor

SuiteSplines.IgaFormation.UnivariatePatchData

SuiteSplines.ImmersedSplines.ClosestExtensionArray

SuiteSplines.ImmersedSplines.ClosestExtensionArray

SuiteSplines.ImmersedSplines.GeneratePerimeter

SuiteSplines.ImmersedSplines.IPerm

SuiteSplines.ImmersedSplines.ImmersedQuadRule

SuiteSplines.KroneckerProducts.BoxProduct

SuiteSplines.KroneckerProducts.KroneckerProduct

SuiteSplines.SortedSequences.IncreasingRange

SuiteSplines.SortedSequences.IncreasingSequence

SuiteSplines.SortedSequences.IncreasingVector

SuiteSplines.SortedSequences.Interval

SuiteSplines.SortedSequences.NonDecreasingSequence

SuiteSplines.SortedSequences.NonDecreasingVector

SuiteSplines.SortedSequences.SortedSequence

SuiteSplines.SortedSequences.Unique

SuiteSplines.SpecialSpaces.Domain

SuiteSplines.SpecialSpaces.MixedFunctionSpace

SuiteSplines.SpecialSpaces.MixedSplineSpace

SuiteSplines.SpecialSpaces.MixedSplineSpaceConstraints

SuiteSplines.SpecialSpaces.MixedSplineSpaceConstraints

SuiteSplines.SpecialSpaces.MixedSplineSpaceConstraints

SuiteSplines.SpecialSpaces.Partition

SuiteSplines.SpecialSpaces.Partition

SuiteSplines.SpecialSpaces.Partition

SuiteSplines.SpecialSpaces.RaviartThomas

SuiteSplines.SpecialSpaces.RaviartThomas

SuiteSplines.SpecialSpaces.ScalarSplineSpace

SuiteSplines.SpecialSpaces.ScalarSplineSpace

SuiteSplines.SpecialSpaces.ScalarSplineSpaceConstraints

SuiteSplines.SpecialSpaces.ScalarSplineSpaceConstraints

SuiteSplines.SpecialSpaces.ScalarSplineSpaceConstraints

SuiteSplines.SpecialSpaces.TaylorHood

SuiteSplines.SpecialSpaces.TaylorHood

SuiteSplines.SpecialSpaces.UnivariateSplineSpaceConstraints

SuiteSplines.SpecialSpaces.UnivariateSplineSpaceConstraints

SuiteSplines.SpecialSpaces.VectorFunctionSpace

SuiteSplines.SpecialSpaces.VectorSplineSpace

SuiteSplines.SpecialSpaces.VectorSplineSpaceConstraints

SuiteSplines.SpecialSpaces.VectorSplineSpaceConstraints

SuiteSplines.SpecialSpaces.VectorSplineSpaceConstraints

SuiteSplines.TensorProductBsplines.TPEvaluationCache

SuiteSplines.TensorProductBsplines.TensorProductBspline

SuiteSplines.UnivariateSplines.Bspline

SuiteSplines.UnivariateSplines.BsplineBasis

SuiteSplines.UnivariateSplines.BsplineEvaluationCache

SuiteSplines.UnivariateSplines.BsplineSupport

SuiteSplines.UnivariateSplines.GeneralizedGaussrule

SuiteSplines.UnivariateSplines.KnotSpanIndices

SuiteSplines.UnivariateSplines.KnotVector

SuiteSplines.UnivariateSplines.SpanIndex

SuiteSplines.UnivariateSplines.SplineSpace

SuiteSplines.UnivariateSplines.Support

SuiteSplines.UnivariateSplines.WeightedQuadrule

Base.adjoint

Base.adjoint

Base.collect

Base.eltype

Base.eltype

Base.getindex

Base.getindex

Base.getindex

Base.insert!

Base.inv

Base.inv

Base.iterate

Base.iterate

Base.iterate

Base.kron

Base.length

Base.length

Base.length

Base.transpose

Base.transpose

LinearAlgebra.det

LinearAlgebra.det

LinearAlgebra.eigen

LinearAlgebra.isposdef

LinearAlgebra.issymmetric

LinearAlgebra.lmul!

LinearAlgebra.logdet

LinearAlgebra.nullspace

LinearAlgebra.nullspace

LinearAlgebra.rmul!

LinearAlgebra.tr

SuiteSplines.AbstractMappings.l2_error

SuiteSplines.BezierBernsteinMethods.Binomial

SuiteSplines.BezierBernsteinMethods.b2c

SuiteSplines.BezierBernsteinMethods.barycentriccoords

SuiteSplines.BezierBernsteinMethods.bernstein

SuiteSplines.BezierBernsteinMethods.bernsteinfuns

SuiteSplines.BezierBernsteinMethods.cartesiancoords

SuiteSplines.BezierBernsteinMethods.decasteljau!

SuiteSplines.BezierBernsteinMethods.degree_elevation_operator

SuiteSplines.BezierBernsteinMethods.dimension

SuiteSplines.BezierBernsteinMethods.dimsplinespace

SuiteSplines.BezierBernsteinMethods.grevillepoint

SuiteSplines.BezierBernsteinMethods.jacobian

SuiteSplines.BezierBernsteinMethods.kformbasisconv

SuiteSplines.BezierBernsteinMethods.linindex

SuiteSplines.BezierBernsteinMethods.mass_matrix

SuiteSplines.BezierBernsteinMethods.multiindex

SuiteSplines.BezierBernsteinMethods.step_decasteljau!

SuiteSplines.BezierBernsteinMethods.subsets

SuiteSplines.BezierBernsteinMethods.vol

SuiteSplines.IgaBase.boundary

SuiteSplines.IgaBase.custom_replace_ref_begin_end!

SuiteSplines.IgaBase.dimension

SuiteSplines.IgaBase.dimension

SuiteSplines.IgaBase.dimension

SuiteSplines.IgaBase.dimension

SuiteSplines.IgaBase.dimension

SuiteSplines.IgaBase.dimension

SuiteSplines.IgaBase.order

SuiteSplines.IgaBase.refine

SuiteSplines.IgaBase.refine

SuiteSplines.IgaFormation.QuadraturePoints

SuiteSplines.IgaFormation.QuadraturePoints

SuiteSplines.IgaFormation.QuadratureWeights

SuiteSplines.IgaFormation.QuadratureWeights

SuiteSplines.IgaFormation.TestFunctions

SuiteSplines.IgaFormation.TestFunctions

SuiteSplines.IgaFormation.TestIndices

SuiteSplines.IgaFormation.TestIndices

SuiteSplines.IgaFormation.TrialFunctions

SuiteSplines.IgaFormation.TrialFunctions

SuiteSplines.IgaFormation.TrialIndices

SuiteSplines.IgaFormation.TrialIndices

SuiteSplines.IgaFormation.coface

SuiteSplines.IgaFormation.pre_allocate_arrays

SuiteSplines.IgaFormation.pre_allocate_arrays

SuiteSplines.IgaFormation.pre_allocate_arrays

SuiteSplines.IgaFormation.pre_allocate_arrays

SuiteSplines.IgaFormation.sumfact!

SuiteSplines.IgaFormation.sumfact!

SuiteSplines.IgaFormation.sumfact!

SuiteSplines.IgaFormation.sumfact!

SuiteSplines.ImmersedSplines.active_splines

SuiteSplines.ImmersedSplines.active_splines

SuiteSplines.ImmersedSplines.active_splines

SuiteSplines.ImmersedSplines.compute_extension_coefficients

SuiteSplines.ImmersedSplines.element_is_inside

SuiteSplines.ImmersedSplines.find_extension_indices

SuiteSplines.ImmersedSplines.find_first_active_element

SuiteSplines.ImmersedSplines.point_is_inside

SuiteSplines.ImmersedSplines.spline_extension_operator

SuiteSplines.KroneckerProducts.box

SuiteSplines.KroneckerProducts.issquare

SuiteSplines.SortedSequences.construct_vector

SuiteSplines.SortedSequences.deconstruct_vector

SuiteSplines.SortedSequences.global_insert

SuiteSplines.SpecialSpaces.dimensions

SuiteSplines.SpecialSpaces.dimensions

SuiteSplines.SpecialSpaces.dimensions

SuiteSplines.SpecialSpaces.dimensions

SuiteSplines.SpecialSpaces.dimensions

SuiteSplines.SpecialSpaces.dimensions

SuiteSplines.SpecialSpaces.extraction_operator

SuiteSplines.SpecialSpaces.extraction_operator

SuiteSplines.SpecialSpaces.extraction_operators

SuiteSplines.SpecialSpaces.extraction_operators

SuiteSplines.SpecialSpaces.getcoeffs

SuiteSplines.SpecialSpaces.indices

SuiteSplines.SpecialSpaces.indices

SuiteSplines.SpecialSpaces.indices

SuiteSplines.SpecialSpaces.indices

SuiteSplines.SpecialSpaces.indices

SuiteSplines.SpecialSpaces.left_constraint!

SuiteSplines.SpecialSpaces.periodic_constraint!

SuiteSplines.SpecialSpaces.right_constraint!

SuiteSplines.SpecialSpaces.setcoeffs!

SuiteSplines.SpecialSpaces.setcoeffs!

SuiteSplines.SpecialSpaces.setcoeffs!

SuiteSplines.SpecialSpaces.setcoeffs!

SuiteSplines.UnivariateSplines.approximate_collocation_inverse

SuiteSplines.UnivariateSplines.approximate_l2_inverse

SuiteSplines.UnivariateSplines.bezier_extraction_operator

SuiteSplines.UnivariateSplines.bspline_integral_value

SuiteSplines.UnivariateSplines.bspline_interpolation_matrix

SuiteSplines.UnivariateSplines.bsplinebasisfuns

SuiteSplines.UnivariateSplines.ders_bspline_interpolation_matrix

SuiteSplines.UnivariateSplines.dersbsplinebasisfuns

SuiteSplines.UnivariateSplines.dimsplinespace

SuiteSplines.UnivariateSplines.distribute_points

SuiteSplines.UnivariateSplines.extraction_operator

SuiteSplines.UnivariateSplines.findspan

SuiteSplines.UnivariateSplines.grevillepoints

SuiteSplines.UnivariateSplines.h_refinement_operator!

SuiteSplines.UnivariateSplines.normalization_weights

SuiteSplines.UnivariateSplines.nquadpoints

SuiteSplines.UnivariateSplines.num_elements

SuiteSplines.UnivariateSplines.onebasisfuneval

SuiteSplines.UnivariateSplines.refinement_operator

SuiteSplines.UnivariateSplines.solve_leastnorm_without_constraints

SuiteSplines.UnivariateSplines.table_required_points

SuiteSplines.UnivariateSplines.two_scale_operator

SuiteSplines.UnivariateSplines.unit_integral_rescaling

SuiteSplines.UnivariateSplines.wquadweights!

SuiteSplines.@suitesplines_reexport

SuiteSplines.AbstractMappings.@evaluate

SuiteSplines.AbstractMappings.@evaluate!

Base.eltype — Method

Base.eltype(::Type{S}) where {Dim,T<:Integer,S<:VectorSplineSpaceConstraints{Dim,<:Any,T}}

Return element type of VectorSplineSpaceConstraints collection.

Base.eltype — Method

Base.eltype(::Type{S}) where {T<:Integer,S<:ScalarSplineSpaceConstraints{<:Any,T}}

Return element type of ScalarSplineSpaceConstraints collection.

Base.getindex — Method

Base.getindex(V::S, i::Int64) where {S<:VectorFunctionSpace}

Return ith vector function space component.

Base.getindex — Method

Base.getindex(C::ScalarSplineSpaceConstraints, i::T) where {T<:Integer}

Return ith component constraints in a ScalarSplineSpaceConstraints

Base.getindex — Method

Base.getindex(C::ScalarSplineSpaceConstraints, i::T) where {T<:Integer}

Return ith component constraints in a VectorSplineSpaceConstraints

Base.iterate — Method

Base.iterate(C::ScalarSplineSpaceConstraints)
Base.iterate(C::ScalarSplineSpaceConstraints, i::T) where {T<:Integer} = iterate(C.data, i)

Iterate of the collection of UnivariateSplineSpaceConstraints stored in ScalarSplineSpaceConstraints.

Base.iterate — Method

Base.iterate(C::VectorSplineSpaceConstraints)
Base.iterate(C::VectorSplineSpaceConstraints, i::T) where {T<:Integer} = iterate(C.data, i)

Iterate of the collection of ScalarSplineSpaceConstraints stored in VectorSplineSpaceConstraints.

Base.iterate — Method

Base.iterate(V::S) where {S<:VectorFunctionSpace} = iterate(parent(V))
Base.iterate(V::S, i::Int) where {S<:VectorFunctionSpace} = iterate(parent(V), i)

Iterate over vector function space components.

Base.length — Method

Base.length(V::S) where {S<:VectorFunctionSpace}

Return the number of vector function space components.

Base.length — Method

Base.length(C::ScalarSplineSpaceConstraints{Dim}) where {Dim}

Return number of dimensions Dim.

Base.length — Method

Base.length(C::VectorSplineSpaceConstraints{<:Any,Codim}) where {Codim}

Return number of dimensions Dim.

SuiteSplines.IgaBase.dimension — Method

dimensions(V::ScalarSplineSpace)

Return the dimension of a ScalarSplineSpace.

SuiteSplines.IgaBase.dimension — Method

IgaBase.dimension(V::T) where {T<:MixedFunctionSpace}

Return the dimension of a mixed function space.

SuiteSplines.IgaBase.dimension — Method

IgaBase.dimension(V::T) where {T<:VectorFunctionSpace}

Return the dimension of a vector function space.

SuiteSplines.IgaBase.dimension — Method

IgaBase.dimension(V::T, k::Int) where {T<:VectorFunctionSpace}

Return the dimension of the ith component of a vector function space.

SuiteSplines.IgaBase.dimension — Method

IgaBase.dimension(V::T, field::Symbol, i::Int64) where {T<:MixedFunctionSpace}

Return the dimension of the ith component of a vector function space field in a mixed space V.

SuiteSplines.IgaBase.dimension — Method

IgaBase.dimension(V::T, field::Symbol) where {T<:MixedFunctionSpace}

Return the dimension of a scalar function space field in a mixed space V.

SuiteSplines.SpecialSpaces.dimensions — Method

dimensions(V::ScalarSplineSpace)

Return the dimension of a ScalarSplineSpace in each tensor-product direction as a tuple.

SuiteSplines.SpecialSpaces.dimensions — Method

dimensions(V::T) where {T<:MixedFunctionSpace}

Return the dimension in each tensor-product direction of a mixed function space as a tuple.

SuiteSplines.SpecialSpaces.dimensions — Method

dimensions(V::T) where {T<:VectorFunctionSpace}

Return the dimension in each tensor-product direction of a vector function space as a tuple.

SuiteSplines.SpecialSpaces.dimensions — Method

dimensions(V::T, i::Int64) where {T<:VectorFunctionSpace}

Return the dimension in each tensor-product direction of the ith component of a vector function space as a tuple.

SuiteSplines.SpecialSpaces.dimensions — Method

dimensions(V::T, field::Symbol, i::Int64) where {T<:MixedFunctionSpace}

Return the dimension in each tensor-product direction of the ith component of a vector function space field in the mixed space V as a tuple.

SuiteSplines.SpecialSpaces.dimensions — Method

dimensions(V::T, field::Symbol, i::Int64) where {T<:MixedFunctionSpace}

Return the dimension in each tensor-product direction of the vector function space field in the mixed space V as a tuple.

SuiteSplines.SpecialSpaces.extraction_operator — Method

extraction_operator(S::ScalarSplineSpace; sparse::Bool=false)

Return the extraction operator for a scalar function space.

If sparse is true, return a sparse matrix.

SuiteSplines.SpecialSpaces.extraction_operator — Method

extraction_operator(S::T, field::Symbol; sparse::Bool=false) where {T<:MixedFunctionSpace}

Return the extraction operator for a scalar function space field in mixed function space S.

If sparse is true, return a sparse matrix.

SuiteSplines.SpecialSpaces.extraction_operators — Method

extraction_operators(S::T; sparse::Bool=false) where {T<:VectorFunctionSpace}

Return the extraction operator for a vector function space.

If sparse is true, return a tuple of sparse matrices.

SuiteSplines.SpecialSpaces.extraction_operators — Method

extraction_operators(S::T, field::Symbol; sparse::Bool=false) where {T<:MixedFunctionSpace}

Return the extraction operators for a vector function space field in mixed function space S in a tuple.

If sparse is true, return a tuple of sparse matrices.

SuiteSplines.SpecialSpaces.getcoeffs — Method

getcoeffs(f::Field)
getcoeffs(f::GeometricMapping)

Return a vector of vertically concatenated (tensor-product Bspline) mapping coefficients.

SuiteSplines.SpecialSpaces.indices — Method

indices(V::ScalarSplineSpace)

Return linear indices for a scalar function space.

SuiteSplines.SpecialSpaces.indices — Method

indices(space::S) where {S<:VectorFunctionSpace}

Return linear indices for a vector function space.

SuiteSplines.SpecialSpaces.indices — Method

indices(space::S, i::Int) where {S<:VectorFunctionSpace}

Return linear indices for ith component of a vector function space.

SuiteSplines.SpecialSpaces.indices — Method

indices(space::S, field::Symbol, i::Int) where {S<:MixedFunctionSpace}

Return linear indices for ith component of a vector function space field in mixed function space space.

SuiteSplines.SpecialSpaces.indices — Method

indices(space::S, field::Symbol) where {S<:MixedFunctionSpace}

Return linear indices for a function space field in mixed function space space.

SuiteSplines.SpecialSpaces.left_constraint! — Method

left_constraint!(C::ScalarSplineSpaceConstraints; c::Vector{Int}=Int[1], dim::Int)

Push c to vector in left field of UnivariateSplineSpaceConstraints stored at index dim in ScalarSplineSpaceConstraints.

SuiteSplines.SpecialSpaces.periodic_constraint! — Method

periodic_constraint!(C::ScalarSplineSpaceConstraints; c::Vector{Int}, dim::Int)

Push c to vector in periodic field of UnivariateSplineSpaceConstraints stored at index dim in ScalarSplineSpaceConstraints.

SuiteSplines.SpecialSpaces.right_constraint! — Method

right_constraint!(C::ScalarSplineSpaceConstraints; c::Vector{Int}=Int[1], dim::Int)

Push c to vector in right field of UnivariateSplineSpaceConstraints stored at index dim in ScalarSplineSpaceConstraints.

SuiteSplines.SpecialSpaces.setcoeffs! — Method

setcoeffs!(f::M, v::Vector{T}, k::Int=1, slice::Base.UnitRange{Int}=Base.UnitRange(1,length(v))) where {M<:AbstractMapping,T<:Real}

Set coefficients of a mapping component to a slice of coefficients in a vector.

Arguments:

f: mapping in question

v: vector with coeffs

slice: range of indices of vector v

k: component of field f

SuiteSplines.SpecialSpaces.setcoeffs! — Method

setcoeffs!(f::M, mixedspace::S, field::Symbol, v::Vector{T}) where {M<:AbstractMapping,S<:MixedSplineSpace,T<:Real}

Set coefficients of a mapping defined on a mixedspace.

Arguments:

f: mapping in question

mixedspace: space

field: space field

v: vector with coeffs with length dim(mixedspace)

SuiteSplines.SpecialSpaces.setcoeffs! — Method

setcoeffs!(f::M, scalarspace::S, v::Vector{T}) where {M<:AbstractMapping,S<:ScalarSplineSpace,T<:Real}

Set coefficients of a mapping defined on a scalar space.

Arguments:

f: mapping in question

scalarspace: space

v: vector with coeffs with length dim(scalarspace)

SuiteSplines.SpecialSpaces.setcoeffs! — Method

setcoeffs!(f::M, vectorspace::S, v::Vector{T}) where {M<:AbstractMapping,S<:VectorSplineSpace,T<:Real}

Set coefficients of a mapping defined on a vector space.

Arguments:

f: mapping in question

vectorspace: space

v: vector with coeffs with length dim(vectorspace)

SuiteSplines.AbstractMappings.Field — Method

Field(space::S) where {S<:ScalarSplineSpace}
Field(space::S) where {S<:VectorSplineSpace}
Field(space::S, s::Symbol) where {S<:MixedSplineSpace}

Construct field on some spline spaces.

SuiteSplines.AbstractMappings.GeometricMapping — Method

AbstractMappings.GeometricMapping(domain, args::ScalarScalarSpace; orientation::Int=1)
AbstractMappings.GeometricMapping(domain, args::VectorSplineSpace; orientation::Int=1)
AbstractMappings.GeometricMapping(domain, args::MixedSplineSpace, field::Symbol; orientation::Int=1)

Construct geometric mapping on some spline spaces.

SuiteSplines.SpecialSpaces.Domain — Type

const Domain{Dim,T} = CartesianProduct{Dim,Tuple{Vararg{T,Dim}},Tuple{Vararg{Interval{T},Dim}}}

Alias for CartesianProduct of Interval defining a domain.

SuiteSplines.SpecialSpaces.MixedFunctionSpace — Type

abstract type MixedFunctionSpace{Dim,T}

Mixed spaces like MixedSplineSpace subtype this.

Mixed function spaces are expected to collect components as struct fields.

SuiteSplines.SpecialSpaces.MixedSplineSpace — Type

abstract type MixedSplineSpace{Dim,T} <: MixedFunctionSpace{Dim,T}

Concrete mixed spaces like RaviartThomas subtype this.

SuiteSplines.SpecialSpaces.MixedSplineSpaceConstraints — Type

const MixedSplineSpaceConstraints

Type alias for NamedTuple which serves as a container for mixed space constraints.

SuiteSplines.SpecialSpaces.MixedSplineSpaceConstraints — Method

MixedSplineSpaceConstraints(S::RaviartThomas{Dim}) where {Dim}

Construct a constraints container for RaviartThomas.

SuiteSplines.SpecialSpaces.MixedSplineSpaceConstraints — Method

MixedSplineSpaceConstraints(S::RaviartThomas{Dim}) where {Dim}

Construct a constraints container for TaylorHood.

SuiteSplines.SpecialSpaces.Partition — Type

const Partition{Dim,T} = CartesianProduct{Dim,Tuple{Vararg{T,Dim}},Tuple{Vararg{IncreasingRange{T},Dim}}}

Alias for CartesianProduct defining a partition.

SuiteSplines.SpecialSpaces.Partition — Method

Partition(S::ScalarSplineSpace{Dimt,T})

Return the partition corresponding to a ScalarSplineSpace.

SuiteSplines.SpecialSpaces.Partition — Method

Partition(domain::S, n::NTuple{Dim, Int64}) where {Dim,T<:Real,S<:Domain{Dim,T}}

Return an uniform Partition of Domain with nₖ breakpoints in kth dimension.

SuiteSplines.SpecialSpaces.RaviartThomas — Type

struct RaviartThomas{Dim,T} <: MixedSplineSpace{Dim,T}

Structure preserving pair of spline spaces for velocities and pressure in two and three dimensions.

SuiteSplines.SpecialSpaces.RaviartThomas — Method

RaviartThomas(p::Degree, Δ::Partition{Dim,T}) where {Dim,T<:Real}
RaviartThomas(p::Degree, Δ::Partition{2,T}, C::MixedSplineSpaceConstraints{(:V,:Q)})
RaviartThomas(p::Degree, Δ::Partition{3,T}, C::MixedSplineSpaceConstraints{(:V,:Q)}) where {T<:Real}

Construct a RaviartThomas mixed spline space.

SuiteSplines.SpecialSpaces.ScalarSplineSpace — Type

Type alias for a tensor-product SplineSpace.

SuiteSplines.SpecialSpaces.ScalarSplineSpace — Method

ScalarSplineSpace(S::ScalarSplineSpace{Dim,T}) where {Dim,T<:Real}
ScalarSplineSpace(p::Degree, partition::Partition{Dim,T}) where {Dim,T<:Real}
ScalarSplineSpace(degrees::NTuple{Dim,Degree}, partition::Partition{Dim,T}) where {Dim,T<:Real}
ScalarSplineSpace(degrees::NTuple{Dim,Degree}, partition::Partition{Dim,T}, C::ScalarSplineSpaceConstraints{Dim}) where {Dim,T<:Real}
ScalarSplineSpace(S::ScalarSplineSpace{Dim,T}, C::ScalarSplineSpaceConstraints{Dim}) where {Dim,T<:Real}

Construct a ScalarSplineSpace.

If constructed from another ScalarSplineSpace, discard constraint conditions.

SuiteSplines.SpecialSpaces.ScalarSplineSpaceConstraints — Type

struct ScalarSplineSpaceConstraints{Dim,T}

Container for ScalarSplineSpace constraints.

Field data stores a tuple of UnivariateSplineSpaceConstraints for each tensor-product dimension of ScalarSplineSpace.

SuiteSplines.SpecialSpaces.ScalarSplineSpaceConstraints — Method

ScalarSplineSpaceConstraints(::F) where {Dim,F<:ScalarSplineSpace{Dim}}

Construct a constraints container for a ScalarSplineSpace.

SuiteSplines.SpecialSpaces.ScalarSplineSpaceConstraints — Method

ScalarSplineSpaceConstraints{Dim}() where {Dim}
ScalarSplineSpaceConstraints{Dim,T}() where {Dim,T<:Integer}

Construct default ScalarSplineSpaceConstraints for a ScalarSplineSpace of dimension Dim.

SuiteSplines.SpecialSpaces.TaylorHood — Type

struct TaylorHood{Dim,T} <: MixedSplineSpace{Dim,T}

An inf-sup stable pair of spline spaces for velocities and pressure in two and three dimensions.

SuiteSplines.SpecialSpaces.TaylorHood — Method

TaylorHood(p::Degree, Δ::Partition{Dim,T}) where {Dim,T<:Real}
TaylorHood(p::Degree, Δ::Partition{Dim,T}, C::MixedSplineSpaceConstraints{(:V,:Q)}) where {Dim,T<:Real}

Construct a TaylorHood mixed spline space.

SuiteSplines.SpecialSpaces.UnivariateSplineSpaceConstraints — Type

struct UnivariateSplineSpaceConstraints{T}

Container for SplineSpace constraints, i.e. cl, cr, cp vectors in SplineSpace constructor stored as left, rightand periodic field.

SuiteSplines.SpecialSpaces.UnivariateSplineSpaceConstraints — Method

UnivariateSplineSpaceConstraints()

Construct default UnivariateSplineSpaceConstraints

SuiteSplines.SpecialSpaces.VectorFunctionSpace — Type

abstract type VectorFunctionSpace{Dim,T}

Concrete vector spaces like VectorSplineSpace subtype this.

Vector function spaces are expected to collect components as a tuple in field :V.

SuiteSplines.SpecialSpaces.VectorSplineSpace — Method

VectorSplineSpace(degree::Degree, partition::Partition{Dim,T}) where {Dim,T<:Real}
VectorSplineSpace(degrees::NTuple{Dim,Degree}, partition::Partition{Dim,T}) where {Dim,T<:Real}
VectorSplineSpace(V::VectorSplineSpace{Dim,Codim,T}, C::VectorSplineSpaceConstraints{Dim,Codim}) where {Dim,Codim,T<:Real}
VectorSplineSpace(S::ScalarSplineSpace{Dim,T}) where {Dim,T<:Real}
VectorSplineSpace(V::NTuple{Codim,ScalarSplineSpace{Dim,T}}) where {Dim,Codim,T}
VectorSplineSpace(V::Vararg{ScalarSplineSpace{Dim,T},Codim}) where {Dim,Codim,T}

Construct a VectorSplineSpace.

Constraints on scalar spline space arguments are preserved.

SuiteSplines.SpecialSpaces.VectorSplineSpaceConstraints — Type

struct ScalarSplineSpaceConstraints{Dim,T}

Container for ScalarSplineSpaceConstraints.

Field data stores a tuple of ScalarSplineSpaceConstraints for each component of VectorSplineSpace.

SuiteSplines.SpecialSpaces.VectorSplineSpaceConstraints — Method

VectorSplineSpaceConstraints(::F) where {Dim,Codim,F<:VectorSplineSpace{Dim,Codim}}

Construct a constraints container for a VectorSplineSpace.

SuiteSplines.SpecialSpaces.VectorSplineSpaceConstraints — Method

VectorSplineSpaceConstraints{Dim}() where {Dim}
VectorSplineSpaceConstraints{Dim,Codim}() where {Dim,Codim}
VectorSplineSpaceConstraints(args::Vararg{ScalarSplineSpaceConstraints{Dim},Codim}) where {Dim,Codim}
VectorSplineSpaceConstraints{T}(args::Vararg{ScalarSplineSpaceConstraints{Dim},Codim}) where {Dim,Codim,T}
VectorSplineSpaceConstraints{Dim,Codim,T}() where {Dim,Codim,T<:Integer}

Construct VectorSplineSpaceConstraints for a VectorSplineSpace of dimension Dim and codomain dimension Codim.